Rebase and improved Lab4
