+++ MobileApp/app/src/main/java/com/mobileapp/views/FourthActivity.kt
@@ -1,11 +1,7 @@
-package com.mobileapp
+package com.mobileapp.views
 
 import android.os.Bundle
 import android.util.Log
-import android.view.View
-import android.widget.Button
-import android.widget.EditText
-import android.widget.ImageButton
 import androidx.appcompat.app.AppCompatActivity
 import com.google.gson.GsonBuilder
 import com.google.gson.JsonParser
@@ -14,7 +10,7 @@ import com.google.mlkit.nl.translate.Translation
 import com.google.mlkit.nl.translate.TranslatorOptions
 import com.mobileapp.services.APIService
 import com.mobileapp.databinding.ActivityFourthBinding
-import com.mobileapp.databinding.ActivityMainBinding
+import com.mobileapp.services.APIService
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
