+++ MobileApp/app/src/main/java/com/mobileapp/views/ThirdActivity.kt
@@ -1,10 +1,9 @@
-package com.mobileapp
+package com.mobileapp.views
 
 import android.os.Bundle
-import android.widget.ImageButton
 import androidx.appcompat.app.AppCompatActivity
 import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.recyclerview.widget.RecyclerView
+import com.mobileapp.utils.ListAdapter
 import com.mobileapp.databinding.ActivityThirdBinding
